# control-api/openapi.yaml
openapi: 3.1.0
info:
  title: AI Media Inference Engine
  description: 'Two surfaces on the same server:


    - **Engine `/v1`** — catalog operations, job queue, assets, text chat.

    - **Control `/api/v1`** — GPU profiles, catalog downloads, generate / upscale / LTX, cleanup.


    Mac proxy: `http://127.0.0.1:8090` → GPU `http://192.168.50.100:8090`.

    Interactive docs: `/docs` (Swagger) and `/redoc`.'
  version: 3.0.0
servers:
- url: http://127.0.0.1:8090
  description: Mac proxy (LAN)
- url: http://192.168.50.100:8090
  description: GPU box (direct)
paths:
  /v1/jobs:
    post:
      tags:
      - Engine (/v1)
      summary: Submit a catalog job
      description: Queue an implemented operation (image.generate, video.generate, text.chat, …). Pass
        Idempotency-Key to replay the same accept response. Inputs use asset_id, not filesystem paths.
      operationId: submit_job_v1_jobs_post
      parameters:
      - name: Idempotency-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Idempotency-Key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubmitJobRequest'
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/jobs/{job_id}:
    get:
      tags:
      - Engine (/v1)
      summary: Get job status
      description: 'Public job record: status, phase, progress, artifacts, error.'
      operationId: get_job_v1_jobs__job_id__get
      parameters:
      - name: job_id
        in: path
        required: true
        schema:
          type: string
          title: Job Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/jobs/{job_id}/events:
    get:
      tags:
      - Engine (/v1)
      summary: Job event stream (SSE)
      description: Server-sent events for job.progress, job.completed, job.failed, job.cancelled. Keepalive
        comments every 15s.
      operationId: job_events_v1_jobs__job_id__events_get
      parameters:
      - name: job_id
        in: path
        required: true
        schema:
          type: string
          title: Job Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/jobs/{job_id}/cancel:
    post:
      tags:
      - Engine (/v1)
      summary: Cancel a job
      operationId: cancel_job_v1_jobs__job_id__cancel_post
      parameters:
      - name: job_id
        in: path
        required: true
        schema:
          type: string
          title: Job Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/jobs/{job_id}/retry:
    post:
      tags:
      - Engine (/v1)
      summary: Retry a job
      description: 'strategy: same | new_seed | override | rerun_stage. override may send new parameters;
        rerun_stage needs stage.'
      operationId: retry_job_v1_jobs__job_id__retry_post
      parameters:
      - name: job_id
        in: path
        required: true
        schema:
          type: string
          title: Job Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RetryJobRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /v1/models/{model_id}:
    get:
      tags:
      - Engine (/v1)
      summary: Get one catalog model
      operationId: get_model_v1_models__model_id__get
      parameters:
      - name: model_id
        in: path
        required: true
        schema:
          type: string
          title: Model Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/models/{model_id}/download:
    post:
      tags:
      - Engine (/v1)
      summary: Start model download
      operationId: download_model_v1_models__model_id__download_post
      parameters:
      - name: model_id
        in: path
        required: true
        schema:
          type: string
          title: Model Id
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/models/{model_id}/verify:
    post:
      tags:
      - Engine (/v1)
      summary: Verify model files on disk
      operationId: verify_model_v1_models__model_id__verify_post
      parameters:
      - name: model_id
        in: path
        required: true
        schema:
          type: string
          title: Model Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/models/{model_id}/load:
    post:
      tags:
      - Engine (/v1)
      summary: Load model onto the GPU
      operationId: load_model_v1_models__model_id__load_post
      parameters:
      - name: model_id
        in: path
        required: true
        schema:
          type: string
          title: Model Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/models/{model_id}/unload:
    post:
      tags:
      - Engine (/v1)
      summary: Unload model from the GPU
      operationId: unload_model_v1_models__model_id__unload_post
      parameters:
      - name: model_id
        in: path
        required: true
        schema:
          type: string
          title: Model Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/models/{model_id}/start:
    post:
      tags:
      - Engine (/v1)
      summary: Start model (alias of load)
      operationId: start_model_v1_models__model_id__start_post
      parameters:
      - name: model_id
        in: path
        required: true
        schema:
          type: string
          title: Model Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/models/{model_id}/stop:
    post:
      tags:
      - Engine (/v1)
      summary: Stop model (alias of unload)
      operationId: stop_model_v1_models__model_id__stop_post
      parameters:
      - name: model_id
        in: path
        required: true
        schema:
          type: string
          title: Model Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /v1/text/chat:
    post:
      tags:
      - Engine (/v1)
      summary: Text + vision chat
      description: Synchronous JSON or SSE when stream=true. Starts llama-fast or gemma as needed. Not
        a Comfy job.
      operationId: text_chat_v1_text_chat_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TextChatRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /v1/workflows/{workflow_id}:
    get:
      tags:
      - Engine (/v1)
      summary: Get one workflow
      operationId: get_workflow_v1_workflows__workflow_id__get
      parameters:
      - name: workflow_id
        in: path
        required: true
        schema:
          type: string
          title: Workflow Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/workflows/{workflow_id}/validate:
    post:
      tags:
      - Engine (/v1)
      summary: Validate a workflow binding
      operationId: validate_workflow_v1_workflows__workflow_id__validate_post
      parameters:
      - name: workflow_id
        in: path
        required: true
        schema:
          type: string
          title: Workflow Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/assets:
    post:
      tags:
      - Engine (/v1)
      summary: Upload an asset
      description: Multipart file upload. Returns asset_id for job inputs.
      operationId: upload_asset_v1_assets_post
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_upload_asset_v1_assets_post'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/assets/{asset_id}:
    get:
      tags:
      - Engine (/v1)
      summary: Get asset metadata
      operationId: get_asset_v1_assets__asset_id__get
      parameters:
      - name: asset_id
        in: path
        required: true
        schema:
          type: string
          title: Asset Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - Engine (/v1)
      summary: Delete an asset
      operationId: delete_asset_v1_assets__asset_id__delete
      parameters:
      - name: asset_id
        in: path
        required: true
        schema:
          type: string
          title: Asset Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
    Body_upload_asset_v1_assets_post:
      properties:
        file:
          type: string
          contentMediaType: application/octet-stream
          title: File
      type: object
      required:
      - file
      title: Body_upload_asset_v1_assets_post
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    RetryJobRequest:
      properties:
        strategy:
          type: string
          enum:
          - same
          - new_seed
          - override
          - rerun_stage
          title: Strategy
          default: same
        parameters:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Parameters
        stage:
          anyOf:
          - type: string
          - type: 'null'
          title: Stage
      type: object
      title: RetryJobRequest
    SubmitJobRequest:
      properties:
        operation:
          type: string
          title: Operation
          description: Catalog operation id, e.g. image.generate, video.generate, video.lipsync
        preset:
          type: string
          title: Preset
          description: draft | balanced | master | character_master | …
          default: master
        inputs:
          additionalProperties: true
          type: object
          title: Inputs
          description: Operation inputs. Use {asset_id} objects for files, not filesystem paths.
        parameters:
          additionalProperties: true
          type: object
          title: Parameters
          description: seed, steps, cfg, loras, duration_seconds, …
        client_context:
          additionalProperties: true
          type: object
          title: Client Context
        output:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Output
      type: object
      required:
      - operation
      title: SubmitJobRequest
    TextChatMessage:
      properties:
        role:
          type: string
          title: Role
        content:
          title: Content
      type: object
      required:
      - role
      - content
      title: TextChatMessage
    TextChatRequest:
      properties:
        model:
          type: string
          title: Model
          description: gemma-4-e4b or qwen36-35b-a3b-rq
        messages:
          items:
            $ref: '#/components/schemas/TextChatMessage'
          type: array
          title: Messages
        temperature:
          anyOf:
          - type: number
          - type: 'null'
          title: Temperature
        max_tokens:
          anyOf:
          - type: integer
          - type: 'null'
          title: Max Tokens
        stream:
          type: boolean
          title: Stream
          description: If true, response is text/event-stream
          default: false
      type: object
      required:
      - model
      - messages
      title: TextChatRequest
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...